requirements.txt
eel==0.16.0
google-generativeai==0.3.1
SpeechRecognition==3.10.0
PyAudio==0.2.13